+++ app/controllers/pages_controller.rb
@@ -5,16 +5,4 @@ class PagesController < ApplicationController
     @items = Item.first(3)
   end
 
-  def create_rental
-    @rental = Rental.new
-    @user = current_user
-    @rental.user = @user
-    @rental.item = @item
-    if @rental.save
-      redirect_to rental_path(@rental)
-    else
-      render :new
-    end
-  end
-
 end
